+++ core/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement `serde::Serialize` and `Deserialize` for `Bytes`.
 - Implement `arbitrary::Arbitrary` for `Bytes`.
 - Add `Serializer::serialize()`.
+- Add `Deserializer::deserialize()`.
 
 ### Fixed